move efp to solvent, simplify routines
