+++ psi4/driver/procrouting/scf_proc/CMakeLists.txt
@@ -1,6 +1,5 @@
 
 list(APPEND sources
-  efp
   scf_iterator
 )
