+++ psi4/driver/procrouting/solvent/CMakeLists.txt
@@ -1,5 +1,6 @@
 
 list(APPEND sources
+  efp
   pol_embed
 )
